+++ Data-Structures/Trees/Treests/BinarySearchTreeTests.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using Xunit;
+using Trees;
+using System.Collections.Generic;
+
+namespace Treests
+{
+    public class BinarySearchTreeTests
+    {
+
+    }
+}
